+++ python/case_uco/__init__.py
@@ -3,4 +3,4 @@
 from case_uco.graph import CASEGraph
 
 __all__ = ["CASEGraph"]
-__version__ = "1.8.0"
+__version__ = "1.9.0"